requirements.txt
click==7.1.2
docx2txt==0.8
lxml==4.6.2
Pillow==8.1.0
python-docx==0.8.10
